Rename Par.chooser to flatMap
Whenever you generalize functions like this, take a critical look at
your generalized function when you're finished. Although the function
may have been motivated by some specific use case, the signature and
implementation may have a more general meaning. In this case, chooser is
perhaps no longer the most appropriate name for this operation, which is
actually quite general — it’s a parallel computation that, when run, will
run an initial computation whose result is used to determine a second
computation. Nothing says that this second computation needs to even
exist before the first computation's result is available. It doesn't
need to be stored in a container like List or Map. Perhaps it's being
generated from whole cloth using the result of the first computation.
This function, which comes up often in functional libraries, is usually
called bind or flatMap.