Solve Exercise 10.9, sort of.
Yeah, I guess I cheated, didn't I?

The whole point of the exercise was to operate on an IndexedSeq[Int]
rather than List[Int], so that you could use foldMapV (which could let
you eventually take advantage of parallelism). I was only able to get
the right answer every time with the monad I came up with when turning
the IndexedSeq into a List.