Add an Applicative-descendant version of Monad
Furthermore, we can now make Monad[F] a subtype of Applicative[F] by
providing the default implementation of map2 in terms of flatMap. This
tells us that all monads are applicative functors, and we don’t need to
provide separate Applicative instances for all our data types that are
already monads.

A minimal implementation of Monad must implement unit and override
either flatMap or join and map.